Expanded nick functions for other features
If a function specifically requests a shortnick, and the player
does not have one set, a truncated 4-char of their name is used.

All shortnicks are force capitalized.

Bool added to check if player has shortnick set (i.e. is result from
function a real nick or a truncated name?)